requirements.txt
pyyaml>=6.0
scikit-learn>=1.3.0
dashscope>=1.14.0
Pillow>=10.0.0
pytest>=7.4.0
numpy>=1.24.0
